% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{get_dates}
\usage{
get_dates(fifa_games)
}
\arguments{
\item{fifa_games}{\itemize{
\item should be df containing cols "game", "full_string", "code", from sfscrape::get_games_codes
}}
}
\value{
df containing the dates for all games in input fifa_games df
}
\description{
Gets all dates listed on sofifa for player stat updates, as well as associated codes for the urls
}
\examples{

fifa_games <- sfscrape::get_game_codes()
dates <- sfscrape::get_dates(fifa_games)

}
