% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_game_codes.R
\name{get_game_codes}
\alias{get_game_codes}
\title{get_game_codes}
\usage{
get_game_codes()
}
\value{
data.frame
}
\description{
Get base codes for each FIFA game.
}
\examples{
sfscrape::get_game_codes()

}
