% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_leagues.R
\name{get_leagues}
\alias{get_leagues}
\title{get_leagues}
\usage{
get_leagues(game_code)
}
\arguments{
\item{game_code}{\itemize{
\item should contain the code for a given fifa game, from sfscrape::get_game_codes
}}
}
\value{
df containing all leagues in FIFA
}
\description{
gets all available leagues for a given year, from the dropdown in https://sofifa.com/teams
}
\examples{

sfscrape::get_leagues(220006)

# to get all games (requires purrr and dplyr installed)
fifa_games <- sfscrape::get_game_codes()
codes <- fifa_games$code
names(codes) <- fifa_games$game
all_leauges <- purrr::map_dfr(codes, get_leagues, .id = "game")

}
