% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_player_data.R
\name{get_player_data}
\alias{get_player_data}
\title{get_player_data}
\usage{
get_player_data(player_code)
}
\arguments{
\item{player_code}{Code associated with a player. Get list of player codes for a given team / date with sfscrape::get_players. Value should be a url suffix like "/player/210257/ederson-santana-de-moraes/220006/"}
}
\value{
df with cols "attribute" and "rating"
}
\description{
Scrapes player data from https://sofifa.com.
}
\examples{

sfscrape::get_player_data("/player/210257/ederson-santana-de-moraes/220006/")

}
