% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_players.R
\name{get_players}
\alias{get_players}
\title{get_players}
\usage{
get_players(team_code, date_code)
}
\arguments{
\item{team_code}{Full code for the team you want to scrape. From sfscrape::get_teams. E.g.}

\item{date_code}{Code for the date you want to scrape. From sfscrape::get_dates}
}
\value{
data.frame with cols "player_id", "player_name", "player_code"
}
\description{
Lists players for a given team and date
}
\examples{

# sfscrape::get_players("/team/10/manchester_city/", "220006")

}
