% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_teams.R
\name{get_teams}
\alias{get_teams}
\title{get_teams}
\usage{
get_teams(league_id, date_code)
}
\arguments{
\item{league_id}{Code for league, from sfscrape::get_leagues}

\item{date_code}{Code for date, from sfscrape::get_game_codes}
}
\value{
df teams, with cols for "team_name", "team_id", "team_code", "date_code"
}
\description{
Gets all teams in a given league / year
}
\examples{

# returns teams from 13 (Premier League), for date 220006 (FIFA 22)
sfscrape::get_teams(13, 220006)

# returns teams in all premier league season, for all seasons available FIFA 07 onwards
fifa_games <- sfscrape::get_game_codes()
codes <- fifa_games$code
names(codes) <- fifa_games$game
all_teams <- purrr::map_dfr(codes, ~get_teams(13, .x), .id = "game")

}
