% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dates.R
\name{get_dates}
\alias{get_dates}
\title{get_dates}
\usage{
get_dates(games)
}
\arguments{
\item{games}{\itemize{
\item Fifa game/s that you want dates for.
}}
}
\value{
df containing the dates for all specified games.
}
\description{
Gets all dates listed on sofifa for player stat updates, as well as associated codes for the urls
}
\examples{

dates <- sfscrape::get_dates("FIFA 22")

}
